Handle null values in snapshot
By default, the first result of a snapshot that doesn't have `initialData` is usually null. Even if you have already called the stream in `initState`. Hence, checking `result.isExecuting` immediately will throw an error. This commit goes ahead to check if `result` is null first, then goes ahead to check for the events